feat: 'revert to floats for dt'

+++ src/FusionEKF.cpp
@@ -152,14 +152,14 @@ void FusionEKF::ProcessMeasurement(const MeasurementPackage &measurement_pack) {
    ****************************************************************************/
 
   cout << "PREDICT\n";
-  double dt = (measurement_pack.timestamp_ - previous_timestamp_) / 1000000.;  //convert from micros to s
+  float dt = (measurement_pack.timestamp_ - previous_timestamp_) / 1000000.;  //convert from micros to s
   previous_timestamp_ = measurement_pack.timestamp_;
 
   cout << "dt = " << dt << endl;
 
-  double dt_2 = dt * dt;
-  double dt_3 = dt_2 * dt;
-  double dt_4 = dt_3 * dt;
+  float dt_2 = dt * dt;
+  float dt_3 = dt_2 * dt;
+  float dt_4 = dt_3 * dt;
 
   //Modify the F matrix so that the time is integrated
   ekf_.F_(0, 2) = dt;
